requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.4.0
tqdm>=4.65.0
pyyaml>=6.0.1
